+++ src/components/Input/Input.css
@@ -18,7 +18,7 @@
 }
 
 
-.InputContainer {
+.passwordAndButtonEye {
   position: relative;
 }
