+++ src/pages/Login/components/CardLogin/CardLogin.css
@@ -13,3 +13,18 @@
 .inputLogin {
   margin-bottom: 46px;
 }
+
+.ErrorValid {
+  text-align: center;
+  color: white;
+  background: #b24444;
+  font-size: 20px;
+  border: none;
+  font-weight: 600;
+  font-family: 'Inter';
+  border-radius: 12px;
+  padding: 50px;
+  position: absolute;
+  left: 75px;
+  bottom: 20px;
+}
