+++ src/pages/Login/components/CardLogin/CardLogin.js
@@ -1,15 +1,84 @@
 
 import './CardLogin.css';
+import { useState } from 'react';
 import loginLogo from '../../../../assets/loginLogo.svg';
 import Input from '../../../../components/Input/Input';
-// import Button from '../../../../components/Button/Button';
+import Button from '../../../../components/Button/Button';
 
 const CardLogin = () => {
+
+  const emptyLogin = 'Login is Empty';
+  const emptyPassword = 'Password is Empty';
+  const emptyLogPass = 'Login & Password is Empty';
+
+  const [ isPasswordEmpty, setIsPasswordEmpty] = useState(false);
+  const [ isLoginEmpty, setIsLoginEmpty] = useState(false);
+  const [ logPass, setLogPass ] = useState(false);
+
+  const [ login, setLogin ] = useState("");
+  const [ password, setPassword ] = useState("");
+
+  const handleChange = (event) => {
+
+    const {name, value} = event.target;
+
+    if(name === "login"){
+      setLogin(value);
+      setIsLoginEmpty(false);
+      setLogPass(false)
+
+      // isLoginEmpty(false);
+      // setState({ isLoginEmpty: false})
+    }
+
+
+
+    if(name === "password"){
+      setPassword(value);
+      setIsPasswordEmpty(false);
+      setLogPass(false);
+    }
+    
+  };
+
+
+  const handleSubmit = (event) => {
+    event.preventDefault();
+    
+    if(login.trim() === "" && password.trim() === ""){
+      setLogPass(true);
+      setIsLoginEmpty(false);
+      setIsPasswordEmpty(false);
+    }
+
+    else if(login.trim() === ""){ setIsLoginEmpty(true); }
+
+    else if(password.trim() === ""){ setIsPasswordEmpty(true) }
+    
+    const user = { login, password }
+
+    if(user.login === 'admin' && user.password === '12345678'){
+      localStorage.setItem('jwt' , '3cwn4u9do92jsb0cg6v82e1');
+      console.log("test for welcome back - Admin");
+    }
+
+    console.log(user)
+  }
+
+
   return (
     <div className='CardLogin'>
       <img className='loginLogo' src={ loginLogo }  alt='loginLogo' />
-      <Input passwordInput />
-      {/* <Button Login/> */}
+
+      <form onSubmit={handleSubmit}>
+        <Input loginInput  value={login} onChange={handleChange} name='login' />
+          {isLoginEmpty && <p className='ErrorValid'>{emptyLogin}</p> }
+        <Input passwordInput  value={password} onChange={handleChange} name='password'  />
+          {isPasswordEmpty && <p className='ErrorValid'>{emptyPassword}</p> }
+        <Button Login/>
+          {logPass && <p className='ErrorValid'>{emptyLogPass}</p>}
+      </form>
+
     </div>
   )
 }
